refactor: move thread utils
