+++ common/api/common.api
@@ -66,7 +66,3 @@ public final class io/opentelemetry/android/common/internal/features/networkattr
 	public static fun values ()[Lio/opentelemetry/android/common/internal/features/networkattributes/data/NetworkState;
 }
 
-public final class io/opentelemetry/android/common/internal/utils/AndroidUtilsKt {
-	public static final fun getThreadIdCompat (Ljava/lang/Thread;)J
-}
-
