+++ instrumentation/common-api/api/common-api.api
@@ -29,3 +29,7 @@ public abstract interface class io/opentelemetry/android/instrumentation/common/
 	public abstract fun extract (Ljava/lang/Object;)Ljava/lang/String;
 }
 
+public final class io/opentelemetry/android/instrumentation/utils/AndroidUtilsKt {
+	public static final fun getThreadIdCompat (Ljava/lang/Thread;)J
+}
+
